+++ core/bean/src/main/java/com/boge/system/bean/vo/ViewPointTagVO.java
@@ -1,11 +1,10 @@
 package com.boge.system.bean.vo;
 
 import com.boge.core.common.base.model.BaseVO;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
 
@@ -27,12 +26,12 @@ public class ViewPointTagVO extends BaseVO<Long> implements Serializable {
      * 观点id
      */
     @ApiModelProperty("观点id")
-    private Integer pointId;
+    private Long pointId;
 
     /**
      * 标签id
      */
     @ApiModelProperty("标签id")
-    private Integer tagId;
+    private Long tagId;
 }
